Improve basic line diff by collapsing changes
The (brilliant) difference crate outputs Add/Remove/Unchanged, but it
makes more sense for a visual diff tool to consider this in terms of
Replaced/Unchanged.

This change moves the diff calculation to a separate module. The
difference crate is then an implementation detail of this, outputting a
vector of differences of its own types.