Fix topological walk to only walk each node once
I had botched the logic to perform the topological walk of all nodes in
the matrix, resulting in an O((|A||B|)^5) complexity (for a set B of
before lines and a set A of after lines), despite my earlier
calculations showing it was possible in linear complexity. This became
obvious on input A and B approaching 100, where the number of nodes to
enumerate exceeded 10E12.

Now fixed this to properly walk the matrix in topological order (and not
expand the walk by the adjacency of every node walked). The walk now
takes O(3|A||B|+|A|+|B|) (i.e. linear) as my proof showed was necessary.